+++ src/persistence/Elsa.Persistence.EntityFrameworkCore/DbContexts/SqliteContext.cs
@@ -4,7 +4,7 @@ namespace Elsa.Persistence.EntityFrameworkCore.DbContexts
 {
     public class SqliteContext : ElsaContext
     {
-        public SqliteContext(DbContextOptions<ElsaContext> options) : base(options)
+        public SqliteContext(DbContextOptions<SqliteContext> options) : base(options)
         {
         }
     }
